+++ css/style.css
@@ -1061,8 +1061,6 @@ fieldset {
 *Сообщения об ошибке загрузки данных
 */
 .data__error {
-  -ms-flex-preferred-size: 100%;
-      flex-basis: 100%;
   color: darkred;
   text-align: center;
 }
